Amazon-AWS-Modify-Instance-Attribute - move QueryApiExecutor where it suppose to be. Signed off by Tusa Mihai: <tusa.mihai@gmail.com>
