+++ cs-jclouds/src/main/java/io/cloudslang/content/jclouds/actions/network/AttachNetworkInterfaceAction.java
@@ -10,7 +10,7 @@ import io.cloudslang.content.jclouds.entities.constants.Outputs;
 import io.cloudslang.content.jclouds.entities.inputs.CommonInputs;
 import io.cloudslang.content.jclouds.entities.inputs.CustomInputs;
 import io.cloudslang.content.jclouds.entities.inputs.NetworkInputs;
-import io.cloudslang.content.jclouds.execute.queries.QueryApiExecutor;
+import io.cloudslang.content.jclouds.execute.QueryApiExecutor;
 import io.cloudslang.content.jclouds.utils.ExceptionProcessor;
 
 import java.util.Map;
